_react/requirements.txt
astroid
certifi
chardet
corsheaders
defusedxml
Django
django-allauth
django-cors-headers
django-rest-auth
djangorestframework
gunicorn
idna
isort
lazy-object-proxy
mccabe
oauthlib
pylint
python3-openid
pytz
requests
requests-oauthlib
six
sqlparse
typed-ast
urllib3
whitenoise
wrapt
